Download, compile, and cache cmake
The source code is currently downloaded from the official sources. It might be nice at some point to instead host a pre-build version on AWS S3 or something.

Upon downloading, cmake is compiled. This takes a long time, but it only hast to happen once for a project. At the moment, all the cmake stuff is put in with the Python stuff. Maybe later there can be a directory just for cmake, but this way everything is already on the $PATH, and the default cache saving/loading logic just works. The path is `.heroku/Python`, and cmake will put things in `.heroku/Python/bin`, and create `.heroku/Python/Modules`.

Later releases of the app will load the cached version, which simply looks for the existence of the cmake binary, similar to how the buildpack finds the cached Python install, for example.