Export a $PREFIX environment variable for compiled dependencies
My own use of this buildpack is to run an app using libgit2 and pygit2, both of which want to install things, and both of which need a bit of guidance doing so in order to play nice. My venvgit2 package helps them out a bit, but at the moment it finds the right place to put stuff by either looking for the current virtual environment, failing that looking for a $PREFIX, and failing that, using some default.

The default (/usr/bin I think) fails with permission errors on heroku, but throwing it in `$PREFIX`, ie. `.heroku/Python` not only works nicely, but gets cached properly too.

This works for now, but a better solution may be to figure out how other compiled packages like `gevent` "just work", and fix venvgit2 (and possibly `pygit2` and also `libgit2`) to just work just like that.